Move web training path from bluenot.py to bluenow.py
